gltf-loader: support multiple data type sizes for index buffers
Each element in an indices array can be of size
* 1, 2, 4. Commet accounts for the change in size.

+++ include/gltf-loader.h
@@ -79,7 +79,7 @@ struct uvr_gltf_loader_vertex_data {
  */
 struct uvr_gltf_loader_vertex_mesh_data {
 	uint32_t                           firstIndex;
-	uint16_t                           *indexBufferData;
+	uint32_t                           *indexBufferData;
 	uint32_t                           indexBufferDataCount;
 	uint32_t                           indexBufferDataSize;
 	struct uvr_gltf_loader_vertex_data *vertexBufferData;
